+++ django_goodies/tests/app/views.py
@@ -1,49 +1,52 @@
 from django.contrib import messages
+from django.shortcuts import render
 
 from django_goodies import AjaxResponse
 
+from .models import CommonInfoTest
 
-def no_args(request):
+
+def ajax__no_args(request):
     
     return AjaxResponse()
 
 
-def request_only(request):
+def ajax__request_only(request):
     
     return AjaxResponse(request)
 
 
-def data(request):
+def ajax__data(request):
     
     return AjaxResponse(request, {'test': 'test'})
 
 
-def bad_data(request):
+def ajax__bad_data(request):
     
     return AjaxResponse(request, ['test1', 'test2'])
 
 
-def success__true(request):
+def ajax__success__true(request):
     
     return AjaxResponse(request, success=True)
 
 
-def success__false(request):
+def ajax__success__false(request):
     
     return AjaxResponse(request, success=False)
 
 
-def success__dumb(request):
+def ajax__success__dumb(request):
     
     return AjaxResponse(request, success='dumb')
 
 
-def success__data(request):
+def ajax__success__data(request):
     
     return AjaxResponse(request, {'test': 'test'}, success=True)
 
 
-def messages_(request):
+def ajax__messages_(request):
     
     messages.success(request, 'This is a success message.')
     messages.error(request, 'This is an error message.')
@@ -52,8 +55,22 @@ def messages_(request):
     return AjaxResponse(request)
 
 
-def messages__data(request):
+def ajax__messages__data(request):
     
     messages.success(request, 'This is a success message.')
     
     return AjaxResponse(request, {'test': 'test'})
+
+
+def ifperm__render(request):
+    
+    return render(request, 'app/ifperm_test.html', {
+        'obj': CommonInfoTest.objects.first()
+    })
+
+
+def ifnotperm__render(request):
+    
+    return render(request, 'app/ifnotperm_test.html', {
+        'obj': CommonInfoTest.objects.first()
+    })