+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Change Log
 
 * Added AjaxResponse
 * Added GOODIES_COMMON_INFO_REQUIRE_USER_ON_SAVE setting
-* Added object-level permission support (ObjectPermissionsBackend)
+* Added object-level permission support (ObjectPermissionsBackend, ifperm and ifnotperm template tags)
 * Updated CommonInfoMixin to provide default object-level permissions for subclasses, based on ownership
 
 0.3